# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Multi-tenant video platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /folders:
    get:
      operationId: listFolders
      tags: [folders]
      parameters:
        - in: query
          name: parentId
          required: true
          schema:
            type: string
            description: A folder UUID, or root for top-level folders.
      responses:
        "200":
          description: Direct child folders in the requested location
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Folder"
    post:
      operationId: createFolder
      tags: [folders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderInput"
      responses:
        "201":
          description: Folder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderDetail"
  /folders/{folderId}:
    get:
      operationId: getFolder
      tags: [folders]
      parameters:
        - $ref: "#/components/parameters/FolderId"
      responses:
        "200":
          description: Folder detail and ordered root-to-parent ancestors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderDetail"
    patch:
      operationId: updateFolder
      tags: [folders]
      parameters:
        - $ref: "#/components/parameters/FolderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderUpdate"
      responses:
        "200":
          description: Updated folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderDetail"
    delete:
      operationId: deleteFolder
      tags: [folders]
      parameters:
        - $ref: "#/components/parameters/FolderId"
      responses:
        "204":
          description: Empty folder deleted
...
components:
  parameters:
...
    FolderId:
      in: path
      name: folderId
      required: true
      schema:
        type: string
...
  schemas:
...
    FolderAncestor:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
    Folder:
      type: object
      required: [id, parentId, name, childFolderCount, videoCount, createdAt, updatedAt]
      properties:
        id: { type: string }
        parentId: { type: ["string", "null"] }
        name: { type: string }
        childFolderCount: { type: number, minimum: 0 }
        videoCount: { type: number, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    FolderDetail:
      type: object
      required: [id, parentId, name, childFolderCount, videoCount, createdAt, updatedAt, ancestors]
      properties:
        id: { type: string }
        parentId: { type: ["string", "null"] }
        name: { type: string }
        childFolderCount: { type: number, minimum: 0 }
        videoCount: { type: number, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        ancestors:
          type: array
          items:
            $ref: "#/components/schemas/FolderAncestor"
    FolderInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        parentId: { type: ["string", "null"] }
    FolderUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        parentId: { type: ["string", "null"] }
...